+++ CHANGELOG.md
@@ -2,8 +2,6 @@
 
 All notable changes to the "vbecm" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
 ### Added
@@ -11,6 +9,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
   For Automatically update modification, and feel slow operation.
 - Editor context menu: Goto Vbe module
   Goto the module on the vbe form the editor you select.
+- Source folder context menu:
+  - pull all modules
+  - compile vba
 
 ### Changed
 - The output tab name is changed to vbecm from myOutput.
@@ -19,6 +20,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - when a vba is running, vbecm stops to access to an excel book. 
   vbecm tests if a continue command and a pause command are enabled in the VBE.
 - Detect modification for no modification module. Fixed. #5
+- Sometime, vbecm can not detect the target for compile. Fixed.
 
 
 ## [0.0.5] - 2022-09-23
@@ -54,6 +56,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 
 <!-- 
+Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 ### Added
  for new features.
 ### Changed